raspberry-pi/requirements-server.txt
opencv-python
flask
flask-cors
requests
RPi.GPIO
numpy
